test: split qbft test setup
